MON-103295 default mariadb plugins directory is configurable and set by default according to the os (#1421)
* default mariadb plugins directory is configurable and set by default according to the os

* fix previous rebase issue

* no lib mariadb static link

* fix test

REFS:MON-103295

+++ common/inc/com/centreon/common/perfdata.hh
@@ -25,17 +25,17 @@ class perfdata {
   enum data_type { gauge = 0, counter, derive, absolute, automatic };
 
  private:
-  double _critical;
-  double _critical_low;
+  float _critical;
+  float _critical_low;
   bool _critical_mode;
-  double _max;
-  double _min;
+  float _max;
+  float _min;
   std::string _name;
   std::string _unit;
-  double _value;
+  float _value;
   data_type _value_type;
-  double _warning;
-  double _warning_low;
+  float _warning;
+  float _warning_low;
   bool _warning_mode;
 
  public:
@@ -48,30 +48,30 @@ class perfdata {
   perfdata();
   ~perfdata() noexcept = default;
 
-  double critical() const { return _critical; }
-  void critical(double c) { _critical = c; }
-  double critical_low() const { return _critical_low; }
-  void critical_low(double c) { _critical_low = c; }
+  float critical() const { return _critical; }
+  void critical(float c) { _critical = c; }
+  float critical_low() const { return _critical_low; }
+  void critical_low(float c) { _critical_low = c; }
   bool critical_mode() const { return _critical_mode; }
   void critical_mode(bool val) { _critical_mode = val; }
-  double max() const { return _max; }
-  void max(double val) { _max = val; }
-  double min() const { return _min; }
-  void min(double val) { _min = val; }
+  float max() const { return _max; }
+  void max(float val) { _max = val; }
+  float min() const { return _min; }
+  void min(float val) { _min = val; }
   const std::string& name() const { return _name; }
   void name(const std::string&& val) { _name = val; }
   void resize_name(size_t new_size);
   const std::string& unit() const { return _unit; }
   void resize_unit(size_t new_size);
   void unit(const std::string&& val) { _unit = val; }
-  double value() const { return _value; }
-  void value(double val) { _value = val; }
+  float value() const { return _value; }
+  void value(float val) { _value = val; }
   data_type value_type() const { return _value_type; };
   void value_type(data_type val) { _value_type = val; }
-  double warning() const { return _warning; }
-  void warning(double val) { _warning = val; }
-  double warning_low() const { return _warning_low; }
-  void warning_low(double val) { _warning_low = val; }
+  float warning() const { return _warning; }
+  void warning(float val) { _warning = val; }
+  float warning_low() const { return _warning_low; }
+  void warning_low(float val) { _warning_low = val; }
   bool warning_mode() const { return _warning_mode; }
   void warning_mode(bool val) { _warning_mode = val; }
 };
